requirements.txt
discord.py==2.3.2
matplotlib==3.9.2
msgspec==0.18.6